pyproject.toml
[project]
name = "australia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "dash>=3.4.0",
    "dash-leaflet>=1.1.3",
    "dbt-bigquery>=1.11.0",
    "dbt-core>=1.11.2",
    "fastexcel>=0.18.0",
    "fsspec>=2025.12.0",
    "gcsfs>=2025.12.0",
    "geopandas>=1.1.2",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-bigquery-storage>=2.36.0",
    "jupyter-dash>=0.4.2",
    "plotly>=6.5.2",
    "polars>=1.37.1",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "streamlit>=1.53.1",
    "watchdog>=6.0.0",
]
